requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==2.0.4
django-sampledatahelper==0.4.1
djangorestframework==3.8.2
httpie==0.9.9
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.1.0
Pygments==2.2.0
pytz==2018.4
requests==2.18.4
sampledata==0.3.7
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
